Fix an unsigned issue in the availability score calculation
Summary: The number of votes and polls are stored as unsigned ints for the computation, causing an overflow when 2 * votes - polls is negative. This can happen if a node never responded any poll, and cause the score to turn high instead of going negative. The test case was checking the score response against a rise then fall, so it didn't check for negative values and missed the edge case.

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D10827

+++ src/net.cpp
@@ -3271,8 +3271,8 @@ void CNode::AvalancheState::updateAvailabilityScore() {
     uint64_t windowInvCounters = invCounters.exchange(0);
     double previousScore = availabilityScore;
 
-    uint32_t polls = windowInvCounters & std::numeric_limits<uint32_t>::max();
-    uint32_t votes = windowInvCounters >> 32;
+    int64_t polls = windowInvCounters & std::numeric_limits<uint32_t>::max();
+    int64_t votes = windowInvCounters >> 32;
 
     availabilityScore =
         AVALANCHE_STATISTICS_DECAY_FACTOR * (2 * votes - polls) +
